+++ src/main/java/net/sf/jabref/logic/formatter/bibtexfields/UnicodeToLatexFormatter.java
@@ -41,7 +41,7 @@ public class UnicodeToLatexFormatter implements LayoutFormatter, Formatter {
         // Standard symbols
         Set<Character> chars = HTMLUnicodeConversionMaps.UNICODE_LATEX_CONVERSION_MAP.keySet();
         for (Character character : chars) {
-            result = result.replaceAll(character.toString(),
+            result = result.replace(character.toString(),
                     HTMLUnicodeConversionMaps.UNICODE_LATEX_CONVERSION_MAP.get(character));
         }